+++ src/models/champions/champion/info.rs
@@ -3,8 +3,8 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Serialize, Deserialize, PartialEq, Clone)]
 
 pub struct Info {
-    pub attack: i16,
-    pub defense: i16,
-    pub magic: i16,
-    pub difficulty: i16,
+    pub attack: f32,
+    pub defense: f32,
+    pub magic: f32,
+    pub difficulty: f32,
 }
