+++ src/models/items/basic/gold.rs
@@ -2,8 +2,8 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct Gold {
-    pub base: i32,
-    pub total: i32,
-    pub sell: i32,
+    pub base: f32,
+    pub total: f32,
+    pub sell: f32,
     pub purchasable: bool,
 }
